E2E_Video_Processing_System/requirements.txt
# FastAPI backend
fastapi
uvicorn[standard]
python-multipart
# aiofiles
# pydantic
python-dotenv
httpx

# Audio/video processing
ffmpeg-python
pydub

# Transcription engines
openai  # For Whisper API

# Segmentation
scikit-learn
hdbscan
umap-learn
keybert
bertopic
sentence-transformers
nltk



#  utilities
requests
numpy
pandas


# Frontend (React) – use package.json
# You would typically run: npx create-react-app frontend --template typescript
# Then install: axios, tailwindcss, etc.

# Dev tools
black
isort
python-dotenv

